fix(browser): cast Statvfs_t.Bsize on netbsd (uint32)
Bare st.Bavail * st.Bsize failed to compile on netbsd (uint64 * uint32),
which broke the goreleaser matrix 36 min into this morning's v0.9.1
release run. Matches the sibling bsd/openbsd/darwin/linux files.

+++ pkg/browser/diskspace_netbsd.go
@@ -10,5 +10,5 @@ func FreeBytes(dir string) (uint64, error) {
 	if err := unix.Statvfs(dir, &st); err != nil {
 		return 0, err
 	}
-	return st.Bavail * st.Bsize, nil
+	return st.Bavail * uint64(st.Bsize), nil
 }
